% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_harmonized_table.R
\name{check_harmonized_table}
\alias{check_harmonized_table}
\title{Checks trait data harmonization}
\usage{
check_harmonized_table(x)
}
\arguments{
\item{x}{A data frame with harmonized trait data}
}
\value{
An (invisible) boolean indicating whether the data frame is acceptable or not.
}
\description{
Checks that the input data frame has the appropriate format and data for model parameter filling.
}
