% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_harmonized_data.R
\name{check_harmonized_trait}
\alias{check_harmonized_trait}
\alias{check_harmonized_allometry}
\title{Check data harmonization}
\usage{
check_harmonized_trait(x)

check_harmonized_allometry(x)
}
\arguments{
\item{x}{A data frame with harmonized trait data or harmonized allometry data}
}
\value{
An (invisible) logical indicating whether the data frame is acceptable or not.
}
\description{
Functions to check that the input data frame (trait table or allometry table) has the appropriate format and data for model parameter filling.
}
\details{
Function \code{check_harmonized_trait()} checks that the input data frame conforms to the following requirements:
  \itemize{
    \item{Has columns called \code{originalName}, \code{acceptedName}, \code{acceptedNameAuthorship}, \code{family}, \code{genus}, \code{specificEpithet}, and \code{taxonRank},
    as returned by function \code{\link{harmonize_taxonomy_WFO}}.}
    \item{The names of the remaining columns are "Units", "Reference", "Priority" or a valid trait name according to the notation required in \code{\link{HarmonizedTraitDefinition}}.}
    \item{For trait columns, their values conform to the required type in \code{\link{HarmonizedTraitDefinition}}.}
  }
Function \code{check_harmonized_allometry()} checks that the input data frame conforms to the following requirements:
  \itemize{
    \item{Has columns called \code{originalName}, \code{acceptedName}, \code{acceptedNameAuthorship}, \code{family}, \code{genus}, \code{specificEpithet}, and \code{taxonRank},
    as returned by function \code{\link{harmonize_taxonomy_WFO}}.}
    \item{Has columns called \code{"Response"} and \code{"Equation"}. Columns \code{"Priority"} and \code{"Reference"} are recommended.}
    \item{Column \code{"Response"} identifies the response variable.}
    \item{Column \code{"ResponseDescription"} contains a longer description of the response variable.}
    \item{Columns \code{"Predictor1"}, \code{"Predictor2"}, ... identify predictor variables.}
    \item{Columns \code{"PredictorDescription1"}, \code{"PredictorDescription2"}, ... contain longer descriptions of predictor variables.}
    \item{Allometry coefficient columns should be named \code{"a"}, \code{"c"}, \code{"c"}, ..., \code{"f"} and contain numeric values.}
  }
}
\seealso{
\code{\link{harmonize_taxonomy_WFO}}, \code{\link{HarmonizedTraitDefinition}}
}
