% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_medfate_params}
\alias{check_medfate_params}
\title{Checks species parameter tables}
\usage{
check_medfate_params(x, check_consistency = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{A data frame with species parameter data}

\item{check_consistency}{A boolean flag to force checking ecological and physiological consistence of parameter values.}

\item{verbose}{A boolean flag to prompt detailed process information.}
}
\value{
An (invisible) list with two data frames, one with missing values in strict parameters and the other with plant names whose parameters (observed or imputed) do not conform to physiological rules.
}
\description{
Checks that the input data frame has the appropriate format and data for simulations with medfate.
}
\details{

}
\examples{
check_medfate_params(SpParamsES, check_consistency = FALSE)
}
\references{
Bartlett et al. (2016). The correlations and sequence of plant stomatal, hydraulic, and wilting responses to drought. PNAS 113: 13098-13103.
}
