% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spparams.R
\name{check_spparams}
\alias{check_spparams}
\title{Checks species parameter tables}
\usage{
check_spparams(x)
}
\arguments{
\item{x}{A data frame with species parameter data}
}
\value{
An (invisible) boolean indicating whether the data frame is acceptable or not.
}
\description{
Checks that the input data frame has the appropriate format and data for simulations with medfate.
}
