% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_strict.R
\name{complete_strict}
\alias{complete_strict}
\title{Complete strict parameters}
\usage{
complete_strict(SpParams, params = NULL)
}
\arguments{
\item{SpParams}{A species parameter data frame to be completed}

\item{param}{A character vector with strict parameters to be completed. If missing, all strict parameters will be checked for completion.}
}
\value{
A species parameter data frame with imputed strict parameter values
}
\description{
Complete strict parameters
}
