% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_traits.R
\name{fill_harmonized_traits}
\alias{fill_harmonized_traits}
\title{Trait filling from harmonized data}
\usage{
fill_harmonized_traits(
  SpParams,
  harmonized_trait_path,
  progress = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{SpParams}{A species parameter data frame to be filled.}

\item{harmonized_trait_path}{A directory path were RDS files with harmonized trait databases are.}

\item{progress}{A boolean flag to prompt progress.}

\item{verbose}{A boolean flag to prompt detailed process information.}
}
\value{
A modified species parameter data frame
}
\description{
Fills species parameter table with trait data
}
