% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_from_observations.R
\name{fill_inventory_traits}
\alias{fill_inventory_traits}
\title{Fills parameters from inventory data}
\usage{
fill_inventory_traits(
  SpParams,
  x,
  quantile_Hmed = 0.5,
  quantile_Hmax = 0.99,
  quantile_fHDmin = 0.05,
  quantile_fHDmax = 0.95,
  progress = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{A list of \code{\link{forest}} objects, or a data frame with a column called \code{forest}, whose elements are of class \code{\link{forest}}.}

\item{quantile_Hmed}{Quantile for Hmed}

\item{quantile_Hmax}{Quantile for Hmax}

\item{quantile_fHDmin}{Quantile for fHDmin}

\item{quantile_fHDmax}{Quantile for fHDmax}

\item{verbose}{}
}
\description{
Extracts species parameter values from forest inventory data
}
