% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_harmonized_traits.R
\name{fill_medfate_traits}
\alias{fill_medfate_traits}
\title{Trait filling from harmonized data}
\usage{
fill_medfate_traits(
  SpParams,
  harmonized_trait_path,
  parameters = NULL,
  priorization = TRUE,
  erase_previous = FALSE,
  replace_previous = FALSE,
  progress = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{SpParams}{A species parameter data frame to be filled for package medfate.}

\item{harmonized_trait_path}{A directory path were RDS files with harmonized trait databases are.}

\item{parameters}{A string vector of parameters to be populated. If \code{NULL} then all possible medfate parameters are parsed.}

\item{priorization}{A boolean flag to perform priorization of some data sources over others.}

\item{erase_previous}{A boolean flag to indicate that all previous values should be set to NA before populating with new data}

\item{replace_previous}{A boolean flag to indicate that non-missing previous values should be replaced with new data}

\item{progress}{A boolean flag to prompt progress.}

\item{verbose}{A boolean flag to prompt detailed process information.}
}
\value{
A modified data frame of medfate species parameters
}
\description{
Fills species parameter table for medfate with trait data from harmonized data sources
}
\details{
The function processes multiple parameters of medfate SpParams table. It identifies the row to modify by
matching column \code{'AcceptedName'} of SpParams with the column \code{acceptedName} of harmonized trait parameter sources.
If the target taxon is a species, values are taken from those rows in trait_table where species names match.
If the target taxon is a genus, then values are taken from those rows where genus is the same.
If \code{priorization = TRUE} and column \code{priority_column} is available in data sources,
the function will prioritize sources with higher priority first, filling parameters with them before inspecting data sources
of lower priority.
}
\seealso{
\code{\link{harmonize_taxonomy_WFO}}
}
