% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_traits.R
\name{fill_traits}
\alias{fill_traits}
\title{Trait filling}
\usage{
fill_traits(SpParams, harmonized_trait_path, verbose = FALSE)
}
\arguments{
\item{SpParams}{A species parameter data frame to be filled}

\item{harmonized_trait_path}{A directory path were RDS files with harmonized trait databases are}

\item{verbose}{A boolean flag for process information}
}
\value{
A modified species parameter data frame
}
\description{
Fills species parameter table with trait data
}
