% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_harmonized.R
\name{get_trait_data}
\alias{get_trait_data}
\alias{load_harmonized_trait_tables}
\alias{load_harmonized_allometry_tables}
\alias{get_allometry_data}
\alias{get_taxon_data}
\title{Load harmonized trait/allometry data}
\usage{
load_harmonized_trait_tables(
  harmonized_trait_path,
  check = TRUE,
  progress = TRUE
)

load_harmonized_allometry_tables(
  harmonized_allometry_path,
  check = TRUE,
  progress = TRUE
)

get_trait_data(
  harmonized_trait_path,
  trait_name,
  taxon_selection_column = NULL,
  taxon_selection = NULL,
  output_format = "long",
  progress = TRUE,
  verbose = TRUE
)

get_allometry_data(
  harmonized_allometry_path,
  response,
  progress = TRUE,
  verbose = TRUE
)

get_taxon_data(
  harmonized_path,
  accepted_name,
  type = "trait",
  progress = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{harmonized_trait_path}{The path to harmonized trait data files (.rds or .csv format).}

\item{check}{A boolean flag to check harmonization and exclude non-acceptable data.}

\item{progress}{A boolean flag to prompt progress.}

\item{harmonized_allometry_path}{The path to harmonized allometry data files (.rds or .csv format).}

\item{trait_name}{A string of an accepted trait name, according to \code{\link{HarmonizedTraitDefinition}}.}

\item{taxon_selection_column}{String indicating the column (e.g. "originalName", "acceptedName", "genus" or "family") that will be used for taxon selection. By default, all taxa are returned.}

\item{taxon_selection}{String vector with taxon names to perform selection. By default, all taxa are returned.}

\item{output_format}{Either "long" or "wide", to indicate output format for trait columns.}

\item{verbose}{A boolean flag to prompt detailed process information.}

\item{response}{String indicating a response variable for allometric equations.}

\item{harmonized_path}{The path to harmonized trait or allometry data files (.rds or .csv format).}

\item{accepted_name}{String of an accepted taxon name.}

\item{type}{A string with data type, either "trait" or "allometry".}
}
\value{
Function \code{load_harmonized_trait_tables()} returns the list with all harmonized trait data tables.

Function \code{load_harmonized_allometry_tables()} returns the list with all harmonized allometry data tables.

Function \code{get_trait_data()} returns a data frame (in long trait format) with the pooled information for a given trait, or an empty data frame if not found.

Function \code{get_taxon_data()} returns a data frame (in long trait format) with the pooled information (traits or allometries) for a given taxon, or an empty data frame if not found.

Function \code{get_allometry_data()} returns a data frame with allometric equations.
}
\description{
Functions to load harmonized trait data and harmonized allometry data
}
\details{
Functions \code{load_harmonized_trait_tables()} and \code{load_harmonized_allometry_tables()} allows loading all kinds of trait data tables, including those that do not pass harmonization check, if \code{check = FALSE}.
In contrast, functions \code{get_trait_data()}, \code{get_taxon_data()} and \code{get_taxon_trait_means()} only return data from data sets passing harmonization checks (see functions \code{\link{check_harmonized_trait}} and \code{\link{check_harmonized_allometry}}).
}
\examples{
\dontrun{
 # List harmonized trait files
 DB_path <- "~/OneDrive/mcaceres_work/model_development/medfate_parameterization/traits_and_models/"
 harmonized_trait_path <- paste0(DB_path,"data/harmonized_trait_sources")
 list.files(path = harmonized_trait_path)

 # Load all harmonized trait data
 l <- load_harmonized_trait_tables(harmonized_trait_path)
 head(l[[1]])

 # Get data for one specific trait
 get_trait_data(harmonized_trait_path, "SLA")

 # Get trait data for one specific taxon
 get_taxon_data(harmonized_trait_path, "Pinus halepensis")

 # List harmonized allometry files
 harmonized_allometry_path = "~/OneDrive/EMF_datasets/AllometryDatabases/Products/harmonized"
 list.files(path = harmonized_allometry_path)

 # Load all harmonized allometry data
 l <- load_harmonized_allometry_tables(harmonized_allometry_path)
 head(l[[1]])

 # Get allometry data for one specific response
 get_allometry_data(harmonized_allometry_path, "FoliarBiomass")
}
}
\seealso{
\code{\link{check_harmonized_trait}}
}
