% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{HarmonizedTraitDefinition}
\alias{HarmonizedTraitDefinition}
\title{Plant trait definitions}
\format{
\itemize{
  \item{Data frame \code{HarmonizedTraitDefinition} has parameters in rows and columns 'Definition', 'Notation', 'Type' and 'Units'.}
}
}
\description{
A data frame of accepted plant trait definitions.
}
\examples{
data(HarmonizedTraitDefinition)
}
\keyword{data}
