% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_params.R
\encoding{UTF-8}
\name{init_medfate_params}
\alias{init_medfate_params}
\title{Initializes species parameter table}
\usage{
init_medfate_params(
  x,
  accepted_names = NULL,
  fill_taxonomy = TRUE,
  complete_rows = TRUE,
  sort = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{A vector of plant names, either taxon names or arbitrary species group names (in that case, \code{accepted_names} should be supplied).
Alternatively, a data frame with columns as returned by function \code{\link{harmonize_taxonomy_WFO}}.}

\item{accepted_names}{Vector of accepted taxon names of the same length of \code{x}.}

\item{fill_taxonomy}{Boolean flag to indicate that taxonomic information should be filled (retrieved from GBIF using package taxize).}

\item{complete_rows}{Boolean flag to indicate that extra rows should be added for cited species/genera (if \code{fill_taxonomy = TRUE}).}

\item{sort}{Boolean flag to force sorting in ascending order by \code{Name}.}

\item{verbose}{A boolean flag to indicate extra console output.}
}
\value{
A data frame with empty species parameter values suitable for medfate simulations.
The data frame will normally contain more rows than \code{sp_names} because of arguments
\code{fill_taxonomy} and \code{complete_rows} are set to TRUE by default.
}
\description{
Creates an empty plant parameter table for medfate, populating taxonomic information if desired.
}
\details{
Taxonomic information is retrieved using functions in package taxize and GBIF as data source.
If \code{x} is a data frame returned by function \code{\link{harmonize_taxonomy_WFO}}, taxonomy is taken from the input data frame,
and package taxize is used for the taxa above family, only.
}
\examples{
\donttest{
# Simple example with two species
sp_names <- c("Salvia rosmarinus", "Pinus contorta")
init_spparams(sp_names, verbose = TRUE)

# Simple example with three species using synonyms and subspecies
sp_names <- c("Rosmarinus officinalis", "Pinus contorta", "Quercus ilex subsp. ilex")
accepted_names <- c("Salvia rosmarinus", "Pinus contorta", "Quercus ilex subsp. ilex")
init_spparams(sp_names, accepted_names, verbose = TRUE)
}
}
\seealso{
\code{\link[medfate]{SpParamsMED}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
