% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medfatetraits-package.R
\docType{package}
\name{medfatetraits-package}
\alias{medfatetraits-package}
\alias{medfatetraits}
\title{medfatetraits: From plant trait data to model parameters}
\description{
Utility functions to facilitate harmonizing plant trait data bases and defining species parameter tables for medfate.
}
\seealso{
Useful links: \itemize{ \item{
\url{https://emf-creaf.github.io/medfatetraits/index.html}} }
}
\author{
Maintainer: Miquel De Cáceres
\email{miquelcaceres@gmail.com}
[ORCID](https://orcid.org/0000-0001-7132-2080)

Author: Víctor Granda
[ORCID](https://orcid.org/0000-0002-0469-1991)

Author: Adriana Tovar

Author: Antoine Cabon
[ORCID](https://orcid.org/0000-0001-6426-1726)

Author: Nicolas Martin-StPaul
[ORCID](https://orcid.org/0000-0001-7574-0108)

Author: Arsène Druel
[ORCID](https://orcid.org/0000-0002-3938-0085)
}
\keyword{internal}
