% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_allometries.R
\encoding{UTF-8}
\name{populate_allometries}
\alias{populate_allometries}
\alias{populate_tree_allometries}
\alias{populate_shrub_allometries_from_medfuels}
\title{Populate tree species allometries}
\usage{
populate_tree_allometries(
  SpParams,
  allom_table,
  allom_type = "foliarbiomass",
  erase_previous = FALSE
)

populate_shrub_allometries_from_medfuels(
  SpParams,
  coef_mapping,
  sp_params_allom,
  group_params_allom,
  species_groups
)
}
\arguments{
\item{SpParams}{A data frame of medfate species parameters to be populated}

\item{allom_table}{A data frame of allometric parameters in columns and taxonomic entities (species or genus) as row names.}

\item{allom_type}{A string with the type of allometry to be filled, either "foliarbiomass", "barkthickness", "crownwidth" or "crownratio".}

\item{erase_previous}{A boolean flag to indicate that values should be set to NA before populating with data}

\item{coef_mapping}{A named string vector specifying which data column should used to populate each medfate param. Elements are data base columns and names are medfate params.}

\item{sp_params_allom}{A data table of species allometric coefficients (typically from package medfuels)}

\item{group_params_allom}{A data table of group allometric coefficients (typically from package medfuels)}

\item{species_groups}{A data table specifying raunkiaer forms for many species (typically from package medfuels)}
}
\value{
A modified data frame of medfate species parameters
}
\description{
Populates allometric coefficients for tree species and genus of an input parameter table on the basis of their
accepted name.
}
\seealso{
\code{\link{initSpParams}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
