% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_from_observations.R
\encoding{UTF-8}
\name{populate_from_observations}
\alias{populate_from_observations}
\alias{populate_GrowthForm}
\alias{populate_height_params}
\alias{populate_tree_diameterHeight_params}
\title{Populate species parameters from inventory data}
\usage{
populate_GrowthForm(
  SpParams,
  tree_names,
  shrub_names = character(0),
  erase_previous = FALSE,
  fill_fromGenus = FALSE
)

populate_height_params(
  SpParams,
  species_names,
  height_values,
  quantile_Hmed = 0.5,
  quantile_Hmax = 0.99,
  erase_previous = FALSE
)

populate_tree_diameterHeight_params(
  SpParams,
  species_names,
  height_values,
  diameter_values,
  quantile_fHDmin = 0.05,
  quantile_fHDmax = 0.95,
  erase_previous = FALSE
)
}
\arguments{
\item{SpParams}{A data frame of medfate species parameters to be populated}

\item{tree_names}{String vector of translated tree species names}

\item{shrub_names}{String vector of translated shrub species names}

\item{erase_previous}{A boolean flag to indicate that values should be set to NA before populating with data}

\item{fill_fromGenus}{A boolean flag to indicate that genus adscription of species should be used to fill missing values}

\item{species_names}{A string vector of translated species names (see \code{\link{translateSpeciesCodes}})}

\item{height_values}{A numeric vector of plant heights (in cm)}

\item{quantile_Hmed}{Quantile for Hmed}

\item{quantile_Hmax}{Quantile for Hmax}

\item{diameter_values}{A numeric vector of tree diameter at breast height (in cm)}

\item{quantile_fHDmin}{Quantile for fHDmin}

\item{quantile_fHDmax}{Quantile for fHDmax}
}
\value{
A modified data frame of medfate species parameters
}
\description{
Internal functions to populate species parameter values from forest inventory source data
}
\seealso{
\code{\link{initSpParams}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
\keyword{internal}
