% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_traits.R
\encoding{UTF-8}
\name{populate_traits}
\alias{populate_traits}
\title{Populate species parameters from trait data}
\usage{
populate_traits(
  SpParams,
  trait_table,
  trait_mapping,
  taxon_column,
  genus_column = NULL,
  character_traits = FALSE,
  summary_function = "median",
  summary_params = list(na.rm = TRUE),
  scalar_functions = NULL,
  replace_previous = FALSE,
  erase_previous = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{SpParams}{A data frame of medfate species parameters to be populated.}

\item{trait_table}{A data frame with functional traits in columns and plants in rows.}

\item{trait_mapping}{A named string vector specifying which trait data column should be used
to populate each medfate param. Elements are data base columns and names are medfate params.}

\item{taxon_column}{A string identifying the column in \code{trait_table} that identifies taxon names).}

\item{genus_column}{A string identifying the column in \code{trait_table} that identifies genus. If \code{NULL} then genus are taken from first word of taxon names.}

\item{character_traits}{Boolean flag to treat traits as character-valued}

\item{summary_function}{A function to summarize multiple values for the same taxonomic entity. By default,
median values are taken for quantitative traits and the most frequent value is taken for qualitative traits.}

\item{summary_params}{A list of summary function params (by default \code{na.rm=TRUE}).}

\item{scalar_functions}{A named list of scalar functions for traits needing transformation of units or scaling. Names are medfate params.}

\item{replace_previous}{A boolean flag to indicate that non-missing previous values should be replaced with new data}

\item{erase_previous}{A boolean flag to indicate that all previous values should be set to NA before populating with new data}

\item{verbose}{A boolean flag to indicate extra console output}
}
\value{
A modified data frame of medfate species parameters
}
\description{
Internal function to fill medfate species trait parameters from a data table of species traits
}
\details{
Matches column 'AcceptedName' of SpParams with trait parameter sources.
If the target taxon is a species, values are taken from those rows in trait_table where species names match.
If the target taxon is a genus, then values are taken from those rows where genus is the same.
}
\seealso{
\code{\link{init_spparams}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
\keyword{internal}
