% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{SpParamsES}
\alias{SpParamsES}
\alias{SpParamsFR}
\alias{SpParamsUS}
\alias{SpParamsAU}
\alias{SpParamsZM}
\title{Data tables with species parameter definition and values for different countries}
\format{
\itemize{
  \item{Data frames \code{SpParamsES} (for Spain), \code{SpParamsFR} (for France), \code{SpParamsAU} (for Australia), \code{SpParamsUS} (for US) and \code{SpParamsZM} (for Zambia) have species or genus as rows and column names equal to parameter names
  in \code{SpParamsDefinition} (the latter from package \code{medfate}).}
}
}
\description{
A data sets of species parameter tables resulting from existing databases to be used in conjunction
with national forest inventory data.
}
\details{
\code{SpParamsES}, \code{SpParamsFR},  \code{SpParamsAU}, \code{SpParamsUS} and \code{SpParamsZM} are species parameter data frames designed to be used with National Forest Inventories
of Spain, France, Australia and USA, respectively.

Details of the procedures used to obtain the species parameter tables can be found in an article at https://emf-creaf.github.io/traits4models/.
}
\examples{
data(SpParamsES)
data(SpParamsFR)
data(SpParamsUS)
data(SpParamsAU)
data(SpParamsZM)
}
\keyword{data}
