% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_species_codes.R
\encoding{UTF-8}
\name{translate_species_codes}
\alias{translate_species_codes}
\title{Translate species codes}
\usage{
translate_species_codes(x, species_names, species_codes)
}
\arguments{
\item{x}{A vector of forest inventory species codes to be mapped.}

\item{species_names}{A string vector of medfate species names with the same length as \code{codes}.}

\item{species_codes}{A string vector of the forest inventory species codes that correspond to the model species codification (\code{species_names}).
Each string may contain different coma-separated codes in order to map different forest inventory species into a single model species.}
}
\value{
A string vector with the length of \code{x} and translated species names
}
\description{
Maps a vector of species codes from forest inventory data to medfate species names
}
\examples{
data(NFI_SP_mapping)
data(piesMayoresIFN2)

piesMayoresIFN2$Especie

translateSpeciesCodes(x = piesMayoresIFN2$Especie,
                      species_names = NFI_SP_mapping$NFIName,
                      species_codes = NFI_SP_mapping$NFICode)

}
\seealso{
\code{\link{IFN2forest}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
