% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_harmonized_table.R
\name{check_harmonized_table}
\alias{check_harmonized_table}
\title{Checks trait data harmonization}
\usage{
check_harmonized_table(x)
}
\arguments{
\item{x}{A data frame with harmonized trait data}
}
\value{
An (invisible) logical indicating whether the data frame is acceptable or not.
}
\description{
Checks that the input data frame has the appropriate format and data for model parameter filling.
}
\details{
The function checks that the input data frame conforms to the following requirements:
  \itemize{
    \item{Has columns called \code{originalName}, \code{acceptedName}, \code{acceptedNameAuthorship}, \code{family}, \code{genus}, \code{specificEpithet}, and \code{taxonRank},
    as returned by function \code{\link{harmonize_taxonomy_WFO}}.}
    \item{The names of the remaining columns are "Units", "Reference" or a valid trait name according to the notation required in \code{\link{HarmonizedTraitDefinition}}.}
    \item{For trait columns, their values conform to the required type in \code{\link{HarmonizedTraitDefinition}}.}
  }
}
\seealso{
\code{\link{harmonize_taxonomy_WFO}}, \code{\link{HarmonizedTraitDefinition}}
}
