% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_medfate_params}
\alias{check_medfate_params}
\title{Checks species parameter tables}
\usage{
check_medfate_params(x)
}
\arguments{
\item{x}{A data frame with species parameter data}
}
\value{
An (invisible) boolean indicating whether the data frame is acceptable or not.
}
\description{
Checks that the input data frame has the appropriate format and data for simulations with medfate.
}
\details{
The function performs the following checks:
\itemize{
  \item{The input is a data frame.}
  \item{All parameter names defined in \code{\link[medfate]{SpParamsDefinition}} should be listed as columns.}
  \item{Strict parameters should not contain missing values. Strict parameters are defined as such in \code{\link[medfate]{SpParamsDefinition}}.}
}
}
