% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_strict.R
\name{complete_strict}
\alias{complete_strict}
\title{Complete strict parameters}
\usage{
complete_strict(SpParams, params = NULL, progress = TRUE, verbose = FALSE)
}
\arguments{
\item{SpParams}{A species parameter data frame to be completed}

\item{progress}{A boolean flag to prompt progress.}

\item{verbose}{A boolean flag to prompt detailed process information.}

\item{param}{A character vector with strict parameters to be completed. If missing, all strict parameters will be checked for completion.}
}
\value{
A species parameter data frame with imputed strict parameter values
}
\description{
Taxonomy-based imputation of strict parameters in medfate.
}
\details{
IMPORTANT: This function performs imputation of strict parameters on the basis of non-missing values in taxonomically-close taxa. It should be used with care because of the
potential source of errors.
}
