% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_strict.R
\name{complete_strict}
\alias{complete_strict}
\title{Complete strict parameters}
\usage{
complete_strict(SpParams, params = NULL, progress = TRUE, verbose = FALSE)
}
\arguments{
\item{SpParams}{A species parameter data frame to be completed}

\item{progress}{A boolean flag to prompt progress.}

\item{verbose}{A boolean flag to prompt detailed process information.}

\item{param}{A character vector with strict parameters to be completed. If missing, all strict parameters will be checked for completion.}
}
\value{
A species parameter data frame with imputed strict parameter values
}
\description{
Taxonomy-based imputation of strict parameters in medfate.
}
\details{
The function performs imputation of strict parameters on the basis of non-missing values in taxonomically-close taxa.
Strict parameters are defined as such in \code{\link[medfate]{SpParamsDefinition}}.
First, if the taxon with missing values is a subspecies, the function looks for values at the species level. Missing values at the species level are filled with data at the
genus, family, order or group levels, depending on the availability of information. Quantitative values are averaged, whereas for qualitative values the first non-missing value is used.
Beware that imputation can result in errors if the initial amount of information is scarce. Therefore, users should revised the output of this function and correct
potential errors manually, whenever detected.
}
