% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_inventory_traits.R
\name{fill_inventory_traits}
\alias{fill_inventory_traits}
\title{Fills parameters from inventory data}
\usage{
fill_inventory_traits(
  SpParams,
  x,
  quantile_Hmed = 0.5,
  quantile_Hmax = 0.99,
  quantile_fHDmin = 0.05,
  quantile_fHDmax = 0.95,
  progress = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{SpParams}{A species parameter data frame to be filled.}

\item{x}{A list of \code{\link{forest}} objects, or a data frame with a column called \code{forest}, whose elements are of class \code{\link{forest}}.}

\item{quantile_Hmed}{Quantile for Hmed}

\item{quantile_Hmax}{Quantile for Hmax}

\item{quantile_fHDmin}{Quantile for fHDmin}

\item{quantile_fHDmax}{Quantile for fHDmax}

\item{progress}{A boolean flag to prompt progress.}

\item{verbose}{A boolean flag to indicate extra console output.}
}
\value{
A modified data frame of medfate species parameters
}
\description{
Extracts species parameter values from forest inventory data
}
\details{
This function fills information of the species parameter table from the data of the target forest inventory where simulations
are to be conducted. Matching is performed between `Species` of the forest inventory data and `Name` of the species parameter table.
The following information is extracted:
\itemize{
\item{\code{GrowthForm}: Growth form according to the usage in the forest inventory. For example, if the species is cited in \code{treeData} tables but not in \code{shrubData} tables, then growth form will be \code{"Tree"}.}
\item{\code{Hmax}: Maximum tree/shrub height (cm), according to the `Height` column in \code{treeData} or \code{shrubData} and \code{quantile_Hmax} parameter.}
\item{\code{Hmed}: Median tree/shrub height (cm), according to the `Height` column in \code{treeData} or \code{shrubData} and \code{quantile_Hmed} parameter.}
\item{\code{fHDmin, fHDmax}: Minimum or maximum height to diameter ratio for trees, according to the `Height` and `DBH` columns in \code{treeData} and \code{quantile_fHDmin} or \code{quantile_fHDmax} parameters, respectively.}
}
}
