% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_allometry.R
\encoding{UTF-8}
\name{fill_medfate_allometries}
\alias{fill_medfate_allometries}
\title{Fill plant allometries}
\usage{
fill_medfate_allometries(
  SpParams,
  harmonized_allometry_path,
  priorization = TRUE,
  erase_previous = TRUE,
  replace_previous = TRUE,
  progress = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{SpParams}{A data frame of medfate species parameters to be populated}

\item{priorization}{A boolean flag to perform priorization of some data sources over others.}

\item{erase_previous}{A boolean flag to indicate that all previous values should be set to NA before populating with new data}

\item{replace_previous}{A boolean flag to indicate that non-missing previous values should be replaced with new data}

\item{progress}{A boolean flag to prompt progress.}

\item{verbose}{A boolean flag to indicate extra console output.}

\item{harmonized_trait_path}{The path to harmonized trait data files (.rds or .csv format).}
}
\value{
A modified data frame of medfate species parameters
}
\description{
Function to populate allometric coefficients for taxa of an input parameter table on the basis of their
accepted name.
}
\seealso{
\code{\link{init_spparams}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
