% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_harmonized_traits.R
\name{get_trait_data}
\alias{get_trait_data}
\alias{load_harmonized_tables}
\title{Load harmonized trait data}
\usage{
load_harmonized_tables(harmonized_trait_path, progress = TRUE)

get_trait_data(
  harmonized_trait_path,
  trait_name,
  is_numeric = TRUE,
  progress = TRUE
)
}
\arguments{
\item{harmonized_trait_path}{The path to harmonized trait data files (.rds)}

\item{progress}{A boolean flag to prompt progress.}

\item{trait_name}{A string of an accepted trait name}

\item{is_numeric}{A boolean indicating whether the trait is numeric}
}
\value{
Function \code{harmonized_table_list()} returns the list with all harmonized trait data tables.

Function \code{get_trait_data()} returns a data table with the pooled information of a single trait.
}
\description{
Functions to load harmonized trait data
}
\examples{
\dontrun{
 harmonized_trait_path = "~/OneDrive/EMF_datasets/PlantTraitDatabases/Products/harmonized"

 # List of files
 trait_files <- list.files(path = harmonized_trait_path)
 head(trait_files)

 # Load trait data
 l <- load_harmonized_tables(harmonized_trait_path)
 head(l[[1]])

 get_trait_data(harmonized_trait_path, "SLA")
}
}
