% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_harmonized.R
\name{get_trait_data}
\alias{get_trait_data}
\alias{load_harmonized_trait_tables}
\alias{load_harmonized_allometry_tables}
\alias{get_allometry_data}
\alias{get_taxon_data}
\title{Load harmonized trait/allometry data}
\usage{
load_harmonized_trait_tables(
  harmonized_trait_path,
  check = TRUE,
  progress = TRUE
)

load_harmonized_allometry_tables(
  harmonized_allometry_path,
  check = TRUE,
  progress = TRUE
)

get_trait_data(
  harmonized_trait_path,
  trait_name,
  output_format = "long",
  progress = TRUE
)

get_allometry_data(harmonized_allometry_path, response, progress = TRUE)

get_taxon_data(
  harmonized_trait_path,
  accepted_name,
  output_format = "long",
  progress = TRUE
)
}
\arguments{
\item{harmonized_trait_path}{The path to harmonized trait data files (.rds or .csv format).}

\item{check}{A boolean flag to check harmonization and exclude non-acceptable data.}

\item{progress}{A boolean flag to prompt progress.}

\item{harmonized_allometry_path}{The path to harmonized allometry data files (.rds or .csv format).}

\item{trait_name}{A string of an accepted trait name}

\item{output_format}{A string, either "wide" or "long", specifying trait output format}

\item{response}{String indicating a response variable for allometric equations.}

\item{accepted_name}{String of an accepted taxon name.}
}
\value{
Function \code{load_harmonized_trait_tables()} returns the list with all harmonized trait data tables.

Function \code{load_harmonized_allometry_tables()} returns the list with all harmonized allometry data tables.

Function \code{get_trait_data()} returns a data frame with the pooled information of a single trait.

Function \code{get_taxon_data()} returns a data frame with the pooled information for a given taxon.
}
\description{
Functions to load harmonized trait data and harmonized allometry data
}
\details{
Function \code{load_harmonized_trait_tables()} allows loading all kinds of trait data tables, including those that do not pass harmonization check, if \code{check = FALSE}.
In contrast, functions \code{get_trait_data()} and \code{get_taxon_data()} only return data from data sets passing harmonization checks (see function \code{\link{check_harmonized_trait}}).
}
\examples{
\dontrun{
 # List harmonized trait files
 DB_path <- "~/OneDrive/mcaceres_work/model_development/medfate_parameterization/traits_and_models/"
 harmonized_trait_path <- paste0(DB_path,"data/harmonized_trait_sources")
 list.files(path = harmonized_trait_path)

 # Load all harmonized trait data
 l <- load_harmonized_trait_tables(harmonized_trait_path)
 head(l[[1]])

 # Get data for one specific trait
 get_trait_data(harmonized_trait_path, "SLA")

 # Get trait data for one specific taxon
 get_taxon_data(harmonized_trait_path, "Pinus halepensis")

 # List harmonized allometry files
 harmonized_allometry_path = "~/OneDrive/EMF_datasets/AllometryDatabases/Products/harmonized"
 list.files(path = harmonized_allometry_path)

 # Load all harmonized allometry data
 l <- load_harmonized_allometry_tables(harmonized_allometry_path)
 head(l[[1]])

 # Get allometry data for one specific response
 get_allometry_data(harmonized_allometry_path, "FoliarBiomass")
}
}
\seealso{
\code{\link{check_harmonized_trait}}
}
