% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_harmonized.R
\name{get_trait_data}
\alias{get_trait_data}
\alias{load_harmonized_tables}
\alias{get_taxon_data}
\title{Load harmonized trait data}
\usage{
load_harmonized_tables(harmonized_trait_path, progress = TRUE)

get_trait_data(
  harmonized_trait_path,
  trait_name,
  is_numeric = TRUE,
  progress = TRUE
)

get_taxon_data(harmonized_trait_path, accepted_name, progress = TRUE)
}
\arguments{
\item{harmonized_trait_path}{The path to harmonized trait data files (.rds)}

\item{progress}{A boolean flag to prompt progress.}

\item{trait_name}{A string of an accepted trait name}

\item{is_numeric}{A boolean indicating whether the trait is numeric}
}
\value{
Function \code{load_harmonized_tables()} returns the list with all harmonized trait data tables.

Function \code{get_trait_data()} returns a data frame with the pooled information of a single trait.

Function \code{get_taxon_data()} returns a data frame with the pooled information for a given taxon.
}
\description{
Functions to load harmonized trait data
}
\details{
Both functions will add \code{Priority = 1} to those trait data sources where \code{Priority} column is not defined.
}
\examples{
\dontrun{
 harmonized_trait_path = "~/OneDrive/EMF_datasets/PlantTraitDatabases/Products/harmonized"

 # List of files
 trait_files <- list.files(path = harmonized_trait_path)
 head(trait_files)

 # Load trait data
 l <- load_harmonized_tables(harmonized_trait_path)
 head(l[[1]])

 get_trait_data(harmonized_trait_path, "SLA")

 # Load taxon data
 get_taxon_data(harmonized_trait_path, "Pinus halepensis")
}
}
