% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_taxonomy_WFO.R
\name{harmonize_taxonomy_WFO}
\alias{harmonize_taxonomy_WFO}
\title{Harmonizes taxonomy}
\usage{
harmonize_taxonomy_WFO(db, WFO_backbone_file)
}
\arguments{
\item{db}{Data frame to harmonize, with species names in column 'originalName'}

\item{WFO_backbone_file}{Path to file containing the backbone of World Flora Online}
}
\value{
A data frame
}
\description{
Harmonizes plant taxonomy according to World Flora Online
}
