% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{HarmonizedTraitDefinition}
\alias{HarmonizedTraitDefinition}
\title{Plant trait definitions}
\format{
A data frame with parameters in rows and columns:
\itemize{
  \item{\code{Definition}: Trait definition}
  \item{\code{Notation}: Trait notation required for harmonization.}
  \item{\code{Type}: Either 'Numeric' (for quantitative traits) or 'String' (for qualitative traits).}
  \item{\code{Units}: Required units for quantitative traits.}
  \item{\code{EquivalentUnits}: Alternative, but equivalent, unit notation for quantitative traits.}
  \item{\code{MinimumValue}: Minimum value for quantitative traits.}
  \item{\code{MaximumValue}: Maximum value for quantitative traits.}
}
}
\description{
A data frame of accepted plant trait definitions.
}
\examples{
data(HarmonizedTraitDefinition)
}
\keyword{data}
