% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_params.R
\encoding{UTF-8}
\name{init_medfate_params}
\alias{init_medfate_params}
\title{Initializes species parameter table}
\usage{
init_medfate_params(x, complete_rows = TRUE, sort = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{A data frame with columns as returned by function \code{\link{harmonize_taxonomy_WFO}}.}

\item{complete_rows}{Boolean flag to indicate that extra rows should be added for cited species/genera (if \code{fill_taxonomy = TRUE}).}

\item{sort}{Boolean flag to force sorting in ascending order by \code{Name}.}

\item{verbose}{A boolean flag to indicate extra console output.}
}
\value{
A data frame with empty species parameter values suitable for medfate simulations.
The data frame will normally contain more rows if \code{complete_rows} is set to TRUE.
}
\description{
Creates an empty plant parameter table for medfate, populating taxonomic information if desired.
}
\details{
Taxonomy is taken from the input data frame, and internal data obtained from package taxize is used for the taxa above family, only.
}
\seealso{
\code{\link[medfate]{SpParamsMED}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
