% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_spparams.R
\encoding{UTF-8}
\name{init_spparams}
\alias{init_spparams}
\title{Initializes species parameter table}
\usage{
init_spparams(
  sp_names,
  SpParamsDefinition,
  accepted_names = NULL,
  fill_taxonomy = TRUE,
  complete_rows = TRUE,
  sort = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sp_names}{Vector of plant names, either taxon names or arbitrary species group names.}

\item{SpParamsDefinition}{Data frame of species parameter definition from package medfate.}

\item{accepted_names}{Vector of accepted taxon names of the same length of \code{sp_names}.}

\item{fill_taxonomy}{Boolean flag to indicate that taxonomic information should be filled (retrieved from GBIF using package taxize).}

\item{complete_rows}{Boolean flag to indicate that extra rows should be added for cited species/genera (if \code{fill_taxonomy = TRUE}).}

\item{sort}{Boolean flag to force sorting in ascending order by \code{Name}.}

\item{verbose}{A boolean flag to indicate extra console output.}
}
\value{
A data frame with empty species parameter values suitable for medfate simulations.
}
\description{
Creates an empty plant parameter table for medfate, populating taxonomic information if desired.
}
\details{
Taxonomic information is retrieved using functions in package taxize and GBIF as data source.
}
\examples{
\dontrun{
# Load species parameter definition from medfate
data(SpParamsDefinition)

# Simple example with two species
sp_names <- c("Salvia rosmarinus", "Pinus contorta")
init_spparams(sp_names, SpParamsDefinition)

# Simple example with three species using synonyms and subspecies
sp_names <- c("Rosmarinus officinalis", "Pinus contorta", "Quercus ilex subsp. ilex")
accepted_names <- c("Salvia rosmarinus", "Pinus contorta", "Quercus ilex subsp. ilex")
init_spparams(sp_names, SpParamsDefinition, accepted_names)


}
}
\seealso{
\code{\link[medfate]{SpParamsMED}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
