% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits4models-package.R
\docType{package}
\name{traits4models-package}
\alias{traits4models-package}
\alias{traits4models}
\title{traits4models: From plant trait data to model parameters}
\description{
Utility functions to facilitate harmonizing plant trait data bases and defining species parameter tables for process-based models
}
\seealso{
Useful links: \itemize{ \item{
\url{https://emf-creaf.github.io/traits4models/index.html}} }
}
\author{
Maintainer: Miquel De Cáceres
\email{miquelcaceres@gmail.com}
[ORCID](https://orcid.org/0000-0001-7132-2080)

Author: Nicolas Martin-StPaul
[ORCID](https://orcid.org/0000-0001-7574-0108)

Author: Adriana Tovar

Author: Víctor Granda
[ORCID](https://orcid.org/0000-0002-0469-1991)
}
\keyword{internal}
